Did not actually need lifetime parameters to express 'static bound.
When updating for the latest rust nightly I accidentally applied the +'static
bound to the outer `Box<>` not the inner `sprite::Updatable<T>`, which meant
the compiler couldn't properly infer the bounds such that `Some(sprite.clone())` was
assignable.

I got around this by using a lifetime parameter, which was set to `'static` anyways.
This wasn't necessary once I put the lifetime bound in the right spot.