(bump version)
0.23.2 should compile on rust-nightly 2015-01-29

NOTE: this is presently using a fork of rust-sdl2 which relaxes certain
guarantees durring the borrowck'ing phase of compilation.

Namely there is no static guarantee that textures outlive their renderer.
This is not an issue for rust-story, as it creates a single renderer which lives
for the duration of the program.

I intend to fix this in a later commit; though it seems like it will requires large
changes to the architecture of the game ...