(hack) lots of gross casts to get it compiling
There's a few unit conversion problems that have cropped up, chiefly:

* Pixel is defined as i32
  * Represents a point on the screen, can be negative for fun and profit
  * Also can be used to represent a length or size of an object

SDL is now using i32 for points on screen, and u32 for lengths.
I will either have to introduce a new unit type to handle this, or live with the casts.

(There's an additional problem in animated sprites; it's using number of frames (u64) to determine a length (u32)
 that becomes an offset (i32)... which means animations are limited to 4 billion frames.)